requirements.txt
alembic==1.8.1
aniso8601==6.0.0
autopep8==1.6.0
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycodestyle==2.8.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
Werkzeug==2.1.2
zipp==3.8.1
